Initialize all data members in QNetworkAccessCache::CacheableObject
Address the "A constructor must initialize all data members of the
class" warning.

The class should accept initial values as argument of a (possibly
protected) constructor if it expects them to be set by a derived
class.

Add an enum Option and a protected constructor that can be called
by the derived classes with the enum values.
This makes setExpires and setShareable redundant so remove them.

Found by an Axivion scan.

Pick-to: 6.8 6.7 6.5
Task-number: QTBUG-125026
Change-Id: Ia8a2a19469a2c0185b5d2e6b2a0895e897f33f28
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>