Call base class assignment operator in QHttpNetworkRequest::operator=()
Address a "Do call the base class copy constructor or assignment
operator, if you implement a copy constructor or assignment operator
in a derived class." warning found by an Axivion scan.

Call the assignment operator of the base class explicitly in the
assignment operator of the derived class.

Task-number: QTBUG-125026
Pick-to: 6.8 6.7 6.5
Change-Id: I5a76c9ea96275d569b3a036b1e1b791c57d0042f
Reviewed-by: Marc Mutz <marc.mutz@qt.io>