CMake: Warn when qt_find_package is used with Qt target names
qt_find_package is only meant to be used for looking up 3rd party
packages. Using it with Qt packages can cause issues where we record
the Qt packages as 3rd party packages.

When using a developer build, issue a warning when qt_find_package is
called with a provided target that starts with the Qt6:: namespace.

Pick-to: 6.8
Change-Id: Ic2a2d613dbf4fa7e7c51e9b73696bf14e4af866f
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>