Comparison helper macros: support variable number of attributes
Comparison macros already had the possibility to specify attributes.
These attributes were added in front of each of the generated
comparison operators.
The original usecase was to allow QT_ASCII_CAST_WARN macro before
each of the generated operators, which was needed e.g. for QString vs
QByteArray comparison [0].
However, further work on applying the macros within QtCore revealed
that the same macro parameter can be used if the operators need to
be template functions [1]. This usecase was, however, limited to
templates with a single type and no constraints.

In order to support templates with more than one type, as well as
complex template constraints, we need to convert the comparison
helper macros into variadic macros, and make Attributes a variable
parameter.

This allows us to specify more complex templates, as well as
template constraints as attributes.

Due to the implementation of QT_OVERLOADED_MACRO(), which supports
maximum of 9 arguments, we can support up to 7 comma-separated
arguments for the attributes (because two marco arguments are used
for the left and right hand side types of the comparison). We need
to explicitly define all nine overloads of the macro, though.
And because of MSVC incorrectly expanding __VA_ARGS__ as one(!)
argument, we need to use QT_VA_ARGS_EXPAND in these definitions.

[0]: 42b6fdfb523f47ba711138bb299d97823e7c64d2
[1]: 890c270b9b27de5075bda668dbe8117649673b05

Task-number: QTBUG-127095
Change-Id: I90f3c53f7135562989eabefa29e05d47264ca089
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>