QPainter: Skip drawing text if target rect is invalid
If drawText() is passed an empty/invalid target rectangle, we would
still go though the motions of laying out and painting the
text. Output would normally be empty, as expected, since the clipping
would take of it. But for paint devices with limited clipping support,
like QSvgGenerator in 1.2 Tiny mode, the text would wrongly show up.

Add a shortcut to skip painting text in this case.

Fixes: QTBUG-127114
Pick-to: 6.8
Change-Id: Ia7de3e7d2493218fc2a0f6cb0e815285ddceeb1d
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>