QTextOption: Add flag for showing default ignorable characters
The presence of the flag instructs the text shaping engine to not
outright ignore normally non-printable code points, so that if the font
has glyphs for them, they will be rendered. This is useful when editing
plain text whose layout is affected by the presence of such characters
(primarily BiDi control characters), and for layout debugging in
general.

[ChangeLog][QtGui][Text] Added QTextOption::ShowDefaultIgnorables flag.

Change-Id: I610a30603f718254ab8e532083e65252351159f0
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
Reviewed-by: Christian Ehrlicher <ch.ehrlicher@gmx.de>