QJniArray: add API to create an empty array and change values
Add a constructor that takes only a size, and create an empty array
using the New*Array function based on the array's type.

Add a QJniArrayMutableIterator. Dereferencing such an iterator returns a
value wrapper, assigning to which overwrites the entry in the Java
array.

Add a mutable operator[](size_type) to QJniArray, which returns the same
value wrapper type. The limitation of this approach is that we cannot
call member functions on the returned value, ie. `strings[1].toLower()`,
as we cannot overload `operator.()`.

`strings[1]->toLower()`

could work, but it's weird, and it won't write any modifications back
into the array (without additional overhead). For that reason the
wrapper deletes the non-const operator->() overload.

This changes the `reference` typedef of QJniArray to that wrapper type.
Since previously no API used that typedef, this is source compatible for
all practical purposes.

[ChangeLog][Core][QJniArray] QJniArray now provides a mutable iterator
and API to create an empty Java array of a specified size, and to set
individual values in the array. QJniArray::reference is now an alias
to a wrapper type that holds a value.

Change-Id: I841591c910ae426d9d1f4b8a2502cfda2a2a874c
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>