Android: Do not request focus on window touch
Remove focusing QtEditText on window touch so that hardware keyboard
events can still be sent to the application after touching the screen.

Move QtEditText focus request back to showSoftwareKeyboard in
QtInputDelegate, so that we can still show virtual keyboard on
QtEditText when needed.

So that QtInputDelegate still has a valid QtEditText instance to show a
virtual keyboard with, make QtWindow set the QtEditText to the
QtInputDelegate on touch.

Task-number: QTBUG-126187
Change-Id: Iff90b96351d5285249d5bcd19f3caa648628cf14
Reviewed-by: Tinja Paavoseppä <tinja.paavoseppa@qt.io>