Android: Do not request focus on window touch
Remove focusing QtEditText on window touch so that hardware keyboard
events can still be sent to the application after touching the screen.

Move QtEditText focus request back to showSoftwareKeyboard in
QtInputDelegate, so that we can still show virtual keyboard on
QtEditText when needed.

So that QtInputDelegate still has a valid QtEditText instance to show a
virtual keyboard with, add a new QtInputConnectionListener interface
function that gives the currently "active" QtEditText to listeners.

Call the new QtInputConnectionListener interface function when QtWindow
is touched.

Pick-to: 6.8 6.7
Task-number: QTBUG-126187
Change-Id: Iff90b96351d5285249d5bcd19f3caa648628cf14
Reviewed-by: Tinja Paavoseppä <tinja.paavoseppa@qt.io>