Rework QtPrivate::Continuation to get rid of virtual calls and rename it
The Continuation class is a template and polymorphic, which likely
means that it instantiates a new type with a vtable for each
combination of template parameters.

Remove the Sync/Async continuation classes to get rid of the hierarcy.
Instead, encode all the information about the type of the continuation
into the class itself. This requires to add two new members:
* QThreadPool (which was previously a member of an async class)
* QRunnable (async class was previously derived from it).

We cannot really eliminate the need for QThreadPool and QRunnable
without a larger refactoring of the entire QFuture framework, because
QFutureInterface requires a QRunnable in its implementation.
However, this patch already allows us to get rid of all the virtual
functions in the Continuation implementation.

While on it, also remove the Function member and derive from
QtPrivate::CompactStorage<Function> instead.

This results in a significant reduce of a binary size for the user
projects, specially if they have a wide usage of continuations.
Taking tst_qfuture as an extreme example, the binary size on a
release build on macOS is reduced by ~15% - from 3723576 to 3156344
bytes.

The change seems to be BC, because all Continuation usages are from
the inline code in other template classes, so most probably a binary
linked with the older version of Qt will simply keep using the old
implementation.
Still, if the user project creates a shared library that utilizes
continuations, and builds the library with the default visibility
attributes, the private Continuation class gets exposed in the
exported methods. To avoid ODR violations in such cases, simply
rename the class to CompactContinuation.

Fixes: QTBUG-124909
Change-Id: I8c8263dbaf407d922aed2017d659bbc6fd87083d
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>