Comparison helper macros: always require static_assert check in C++20
Commit d292648d0bbac50388dae035dc34782accb3807f introduced
QT_COMPARISON_NOEXCEPT_CHECK macro to suppress static_assert() checks
on QNX and Integrity compilers that have bugs with handling conditional
noexcept on lambdas.
Update the macro to always require static_assert() in C++20 builds,
because we want the compilers to fix the known issues before we update.

Task-number: QTBUG-127095
Change-Id: Ic90cc95330d8e6fe085fdc8c51bbca75a65240fd
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>