Revert "Revert "QSharedPointer: Use new comparison helper macros""
This reverts commit 0ac0b412c754e53d167716f207be9d76a7fe16be.

Reason for revert: the problem with the Integrity compiler is now
solved by not using conditional noexcept in the comparison helper
macros. It should be fine to merge the QSharedPointer changes again.

Some fixes were made to the original patch, to adjust to the
updates in comparison helper macros.

Task-number: QTBUG-120306
Change-Id: I4c6eb141ca650aead62c283874e83c27c6cf79b1
Reviewed-by: Rym Bouabid <rym.bouabid@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>